+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 -------------------
 
 - fix: :meth:`.Table.distinct` now deduplicates rows when ``key`` is a sequence of column names.
+- fix: :class:`.Rank` ranks null values last when ``reverse=True``.
 
 1.14.2 - February 27, 2026
 --------------------------